refactor: 배지 계산하는 로직을 useVisibleBadgeCount 커스텀 훅으로 분리
